+++ src/main/java/gift/front/controller/MemberFrontController.java
@@ -1,6 +1,5 @@
 package gift.front.controller;
 
-import gift.api.option.service.OptionService;
 import gift.api.product.service.ProductService;
 import gift.api.wish.service.WishService;
 import jakarta.servlet.http.Cookie;
@@ -23,13 +22,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class MemberFrontController {
 
     private final ProductService productService;
-    private final OptionService optionService;
     private final WishService wishService;
 
-    public MemberFrontController(ProductService productService, OptionService optionService,
-            WishService wishService) {
+    public MemberFrontController(ProductService productService, WishService wishService) {
         this.productService = productService;
-        this.optionService = optionService;
         this.wishService = wishService;
     }
 
@@ -83,7 +79,7 @@ public class MemberFrontController {
             Model model
     ) {
         model.addAttribute("product", productService.findProductById(id));
-        model.addAttribute("options", optionService.getOptionsByProductId(id));
+        model.addAttribute("options", productService.getOptionsByProductId(id));
 
         return "member/product-detail";
     }
